+++ pytorch_benchmark/benchmark.py
@@ -146,7 +146,7 @@ def measure_repeated_inference_timing(
     t_tot = []
 
     for _ in tqdm(
-        range(num_runs), desc=f"Measuring inference with batch_size={batch_size}"
+        range(num_runs), desc=f"Measuring inference for batch_size={batch_size}"
     ):
         start_on_cpu = time()
         device_sample = transfer_to_device_fn(sample, model_device)
@@ -203,7 +203,8 @@ def measure_energy(
     sample,
     model_device,
     transfer_to_device_fn=torch.Tensor.to,
-    print_details=False,
+    num_runs=100,
+    batch_size: int = None,
     include_transfer_costs=True,
     print_fn=logger.info,
 ):
@@ -254,8 +255,12 @@ def measure_energy(
 
         p_est = PowerEstimator(print_fn=print_fn)
         # index 0 is total energy, index 1 is energy over idle consumption:
-        total_joules = p_est.estimate_fn_power(test_fn)[0] / 1000
-        inference_joules = total_joules
+        meas = []
+        for _ in tqdm(
+            range(num_runs), desc=f"Measuring energy for batch_size={batch_size}"
+        ):
+            meas.append(p_est.estimate_fn_power(test_fn)[0] / 1000)
+        inference_joules = float(np.array(meas).mean())
     except Exception:
         pass
 
@@ -324,34 +329,40 @@ def benchmark(
         results["flops"] = flops
         print_fn(f"Model FLOPs: {flops} ({format_num(flops)})")
 
-    # Measure Allocated Memory
-    if model_device.type == "cuda":
-        pre_mem, post_mem, max_mem = measure_allocated_memory(
-            model, sample, model_device, transfer_to_device_fn, print_details
-        )
-        results["pre_inference_memory"] = pre_mem
-        results["max_inference_memory"] = max_mem
-        results["post_inference_memory"] = post_mem
-        print_fn(
-            f"Allocated GPU memory prior to inference: {pre_mem} ({format_num(pre_mem, bytes=True)})"
-        )
-        print_fn(
-            f"Allocated GPU memory after to inference: {post_mem} ({format_num(post_mem, bytes=True)})"
-        )
-        print_fn(
-            f"Max allocated GPU memory during inference: {max_mem} ({format_num(max_mem, bytes=True)})"
-        )
-    else:
-        logger.warning(
-            "Measurement of allocated memory is only available on CUDA devices"
-        )
-
     # Measure inference timing
+    memory = {}
     timing = {}
     energy = {}
     with torch.no_grad():
         for bs in sorted(set([1, batch_size])):
             s = sample1 if bs == 1 else sample
+
+            # Measure Allocated Memory
+            if model_device.type == "cuda":
+                pre_mem, post_mem, max_mem = measure_allocated_memory(
+                    model, sample, model_device, transfer_to_device_fn, print_details
+                )
+                memory[f"batch_size_{bs}"] = {
+                    "pre_inference_bytes": pre_mem,
+                    "max_inference_bytes": max_mem,
+                    "post_inference_bytes": post_mem,
+                    "pre_inference": format_num(pre_mem, bytes=True),
+                    "max_inference": format_num(max_mem, bytes=True),
+                    "post_inference": format_num(post_mem, bytes=True),
+                }
+                print_fn(
+                    fmt(
+                        {
+                            f"Memory results (batch_size={bs})": memory[
+                                f"batch_size_{bs}"
+                            ]
+                        }
+                    )
+                )
+            else:
+                logger.warning(
+                    "Measurement of allocated memory is only available on CUDA devices"
+                )
 
             # Inference timing
             warm_up(
@@ -380,11 +391,13 @@ def benchmark(
             # Energy measurement
             energy_joules = measure_energy(
                 model,
-                sample,
+                s,
                 model_device,
                 transfer_to_device_fn,
-                print_details,
-                print_fn,
+                num_runs=max(1, num_runs // 10),
+                batch_size=bs,
+                include_transfer_costs=True,
+                print_fn=print_fn,
             )
             if _is_valid(energy_joules):
                 energy_kwh = energy_joules / 3.6e6